refactor to mixin class and use model runner
